The transient write path ships: filling, flushing and sealing
`tb_flush`, `b_conj` and `b_persistent`. `Bytes` across the three runtimes:
2 212 lines when this started and 1 022 now -- 54% of it gone.

THIS WAS THE LAST THING HOLE 5 WAS BLOCKING. A FULL tail is handed to the tree
whole -- it is exactly the leaf the tree wants, so there is nothing to copy --
but a PARTIAL one has to be copied out to its real length first, because a
leaf's length IS its byte count and handing over a half-filled buffer would
claim it holds `TAIL_CAP` bytes. Copying bytes out is what there was nowhere
to put, and `b_copy_range` is now the whole of that.

`write-u8` is the other half of `read-u8` and is what appending actually is:
one byte into the open tail, no allocation and no copy until it fills.

THE PROBE SETS `TAIL_CAP` TO FOUR so the flush boundary is reachable by hand,
and checks the sequence rather than the total: three bytes leave the tree nil,
the fifth flushes `1.2.3.4` into it whole, and sealing joins the partial `5`.
A byte count alone passes with the flush never happening.

It also checks the two things a transient is FOR: sealing invalidates it, and
using a dead one is refused rather than silently reading a structure somebody
else now owns. And that nothing-appended answers the empty byte string rather
than nil, so a caller does not have to test.

`byte` had to be renamed to `bv`. It is a Java keyword and Java has no verbatim
identifier -- Rust would take `r#byte` and C# `@byte` -- so kin refuses it by
name rather than letting two targets spell it one way and one another. That
refusal is `test/reserved.clj`, written this morning for exactly this.

Gates: bin/test 0, conform 26 rows + 13 GC-stress, 0 FAIL, check-kin green over
39 sources, opcov 217/256, threads 305 075 < 307 000, 130 Rust tests under
`--all-targets --features diagnostics`.

+++ runtime/src/builtins.rs
@@ -320,7 +320,7 @@ builtins! {
         let _ = n;
         let (t, x) = (arg(rt, a, 0), arg(rt, a, 1));
         match rt.as_i64(x) {
-            Some(b) => rt.b_conj(t, (b & 0xff) as u8),
+            Some(b) => rt.b_conj(t, (b & 0xff) as u32),
             None => rt.throw_str("ClassCastException", "b-conj! wants an integer"),
         }
     };
